vmscan/swapfile: added likely/unlikely hints around checks for free swap space.
With ZRAM enabled, we should always have swap space available, and these checks
should always return true, so giving the compiler branch prediction hints to
this effect should be a (small) performance win, especially since one of them is
in a codepath described in the comments as "the hottest path in the kernel, for
pagecache-intensive workloads".

If we ever end up disabling ZRAM, this commit should be reverted.